Move MoEComputeStage into stage.py
